+++ lib/request_code.dart
@@ -1,15 +1,17 @@
 import 'dart:async';
-import 'package:flutter/material.dart';
 
-import 'request/authorization_request.dart';
-import 'model/config.dart';
+import 'package:flutter/material.dart';
 import 'package:webview_flutter/webview_flutter.dart';
+
+import 'model/config.dart';
+import 'request/authorization_request.dart';
 
 class RequestCode {
   final Config _config;
   final AuthorizationRequest _authorizationRequest;
-  final _redirectUriHost;
+  final String _redirectUriHost;
   late NavigationDelegate _navigationDelegate;
+  late WebViewCookieManager _cookieManager;
   String? _code;
 
   RequestCode(Config config)
@@ -19,6 +21,7 @@ class RequestCode {
     _navigationDelegate = NavigationDelegate(
       onNavigationRequest: _onNavigationRequest,
     );
+    _cookieManager = WebViewCookieManager();
   }
 
   Future<String?> requestCode() async {
@@ -33,6 +36,13 @@ class RequestCode {
     await controller.setBackgroundColor(Colors.transparent);
     await controller.setUserAgent(_config.userAgent);
     await controller.loadRequest(launchUri);
+    if (_config.onPageFinished != null) {
+      await controller.setNavigationDelegate(
+        NavigationDelegate(
+          onPageFinished: _config.onPageFinished,
+        ),
+      );
+    }
 
     final webView = WebViewWidget(controller: controller);
 
@@ -48,13 +58,17 @@ class RequestCode {
     await _config.navigatorKey.currentState!.push(
       MaterialPageRoute(
         builder: (context) => Scaffold(
-          body: WillPopScope(
-            onWillPop: () async {
+          appBar: _config.appBar,
+          body: PopScope(
+            canPop: false,
+            onPopInvoked: (bool didPop) async {
+              if (didPop) return;
               if (await controller.canGoBack()) {
                 await controller.goBack();
-                return false;
+                return;
               }
-              return true;
+              final NavigatorState navigator = Navigator.of(context);
+              navigator.pop();
             },
             child: SafeArea(
               child: Stack(
@@ -88,7 +102,7 @@ class RequestCode {
   }
 
   Future<void> clearCookies() async {
-    await WebViewCookieManager().clearCookies();
+    await _cookieManager.clearCookies();
   }
 
   String _constructUrlParams() => _mapToQueryParams(