+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## [Unreleased]
 
+## [1.0.1] - 20230215
+
+- Optionally avoid prompting popup on web logout https://github.com/Earlybyte/aad_oauth/pull/277 (by @phildupuis)
+- Fix/inline linter rules to prevent conflict https://github.com/Earlybyte/aad_oauth/pull/279 (by @mikedigit)
+- **Breaking** Update willpopscope to popscope https://github.com/Earlybyte/aad_oauth/pull/291 (by @elijahjohnston) fix: either upgrade Flutter to 3.16 or downgrade AAD_OAUTH to 1.0.0
+- Add login hint for web https://github.com/Earlybyte/aad_oauth/pull/292 (by @elijahjohnston)
+- ✨ feat: add onPageFinished function in Config https://github.com/Earlybyte/aad_oauth/pull/280 (by @tajjacob)
+
+## [1.0.0] - 20231010
+
+- Add ADFS support for mobile applications https://github.com/Earlybyte/aad_oauth/pull/270 (by @AndreasHassing)
+- Add explicit and implicit refresh token https://github.com/Earlybyte/aad_oauth/pull/260 (by @easazade)
+- Plugins: update multiple plugins and ui https://github.com/Earlybyte/aad_oauth/pull/257 (by @tajjacob)
+- Remove client secret from Azure AD B2C config instructions in README https://github.com/Earlybyte/aad_oauth/pull/252 (by @OssiPesonen)
+- Update deps https://github.com/Earlybyte/aad_oauth/pull/245 (by @Zazo032)
+- Authority improvements https://github.com/Earlybyte/aad_oauth/pull/236 (by @Zazo032)
+- Handle pop in the webview https://github.com/Earlybyte/aad_oauth/pull/233 (by @xhoi-xhafaj)
+
 ## [0.4.4] - 20230516
 
 - Add cacheLocation configuration https://github.com/Earlybyte/aad_oauth/pull/220 (by @tigloo)